Transaction format function moved to SDK.transaction
